node-data: move ledger::transaction in its own module
